refactor(discovery): decompose source-sync.ts into focused sub-modules
Closes #270

source-sync.ts (1,916 lines) was the last large monolith in the
discovery domain. Decomposed into 16 focused files under a new
source-sync/ directory, retaining source-sync.ts as a thin re-export
barrel so all existing import paths remain unchanged.

New files
─────────
src/domains/discovery/source-sync/
  types.ts          — shared interfaces (leaf; no sibling imports)
  state.ts          — state I/O, cursor helpers, entry upsert/prune
  fetching.ts       — SSRF-safe fetch wrappers, URL normalization,
                      sitemap/HTML parsing, misc data helpers
  references.ts     — URL → indexed-reference catalog entry builder
  reporting.ts      — non-indexed source classification
  html.ts           — generic sitemap + paginated-HTML sync engines
  index.ts          — orchestrator + synchronizeIndexedSource dispatch

  registries/
    vscode-marketplace.ts  — VS Code Marketplace query/page cursor sync
    clawhub.ts             — ClawHub server-rendered plugin catalog
    npm.ts                 — npm replicate changes-feed sync
    crates.ts              — crates.io paginated search API
    go.ts                  — pkg.go.dev sitemap + HTML sync
    maven.ts               — Maven Central Sonatype search API
    mcp-registry.ts        — MCP registry paginated servers API
    nuget.ts               — NuGet V3 search query service
    packagist.ts           — Packagist paginated search API

Constraints upheld
──────────────────
- All existing import paths intact (barrel re-exports everything)
- sourceSyncInternals test-escape-hatch preserved on barrel
- 776/776 tests pass; 100% statement/branch/function/line coverage
- Lint clean (eslint --max-warnings=0)
- TypeScript strict-mode clean
